pyproject.toml
[project]
name = "mycv"
version = "0.1.0"
description = "CV builder: YAML data + Jinja2 template → HTML/PDF"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "weasyprint>=68.1",
]
